Pasarela de pago (guardar orders y facturas en la base de datos)

+++ database/migrations/2024_04_19_100718_order_product.php
@@ -21,7 +21,8 @@ return new class extends Migration
             $table->primary(['id_order', 'id_product']);
 
             $table->integer('quantity');
-            $table->integer('price');
+            $table->unsignedBigInteger('price');
+
             $table->timestamps();
         });
     }
